port some transformstream tests to w-p-t
