+++ reference-implementation/to-upstream-wpts/transform-stream/general.https.html
@@ -4,6 +4,8 @@
 <script src="/resources/testharnessreport.js"></script>
 <script src="/service-workers/service-worker/resources/test-helpers.sub.js"></script>
 <script src="../resources/test-initializer.js"></script>
+<script src="../resources/test-utils.js"></script>
+<script src="../resources/rs-utils.js"></script>
 
 <script src="general.js"></script>
 <script>
